feat: enhance CompanyController and CompanyService with detailed API responses, error handling, and pagination support

+++ src/modules/company/company.service.ts
@@ -1,29 +1,71 @@
 import { PrismaService } from '@database/PrismaService';
-import { Injectable } from '@nestjs/common';
-import type { UpdateCompanyDto } from './dto/update-company.dto';
-import type { CreateCompanyDto } from './dto/create-company.dto';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ResponseCreateCompanyDto,
+  ResponseDeleteCompanyDto,
+  ResponseGetCompanyDto,
+  ResponsePaginatedCompaniesDto,
+  ResponseUpdateCompanyDto,
+} from '@modules/company/dto/response-company.dto';
+import { QueryPaginationDto } from 'src/dtos/query-pagination.dto';
+import { CreateCompanyDto } from '@modules/company/dto/create-company.dto';
+import { UpdateCompanyDto } from '@modules/company/dto/update-company.dto';
 
 @Injectable()
 export class CompanyService {
   constructor(private prisma: PrismaService) {}
 
-  create(data: CreateCompanyDto) {
+  async create(data: CreateCompanyDto): Promise<ResponseCreateCompanyDto> {
     return this.prisma.company.create({ data });
   }
 
-  findAll() {
-    return this.prisma.company.findMany();
+  async findAll(query: QueryPaginationDto): Promise<ResponsePaginatedCompaniesDto> {
+    const take = Number(query.take) || 10; // max products per page
+    const skip = Number(query.skip) || 1; // page number
+
+    const [companies, count] = await Promise.all([
+      this.prisma.company.findMany({
+        take,
+        skip: (skip - 1) * take,
+      }),
+      this.prisma.company.count(),
+    ]);
+
+    const pages = Math.ceil(count / take);
+
+    return {
+      companies,
+      pages: pages === 0 ? 1 : pages,
+      count,
+    };
   }
 
-  findOne(id: number) {
+  async findOne(id: number): Promise<ResponseGetCompanyDto> {
+    const company = await this.prisma.company.findUnique({ where: { id } });
+    if (!company) {
+      throw new NotFoundException('Empresa não encontrada');
+    }
+
     return this.prisma.company.findUnique({ where: { id } });
   }
 
-  update(id: number, dto: UpdateCompanyDto) {
+  async update(id: number, dto: UpdateCompanyDto): Promise<ResponseUpdateCompanyDto> {
+    const company = await this.prisma.company.findUnique({ where: { id } });
+    if (!company) {
+      throw new NotFoundException('Empresa não encontrada');
+    }
+
     return this.prisma.company.update({ where: { id }, data: dto });
   }
 
-  remove(id: number) {
-    return this.prisma.company.delete({ where: { id } });
+  async remove(id: number): Promise<ResponseDeleteCompanyDto> {
+    const company = await this.prisma.company.findUnique({ where: { id } });
+    if (!company) {
+      throw new NotFoundException('Empresa não encontrada');
+    }
+
+    await this.prisma.company.delete({ where: { id } });
+
+    return { message: 'Empresa excluída com sucesso' };
   }
 }
